Develop

+++ back/src/Controller/GradeController.php
@@ -29,16 +29,18 @@ class GradeController extends AbstractController
 
         $form->handleRequest($request);
 
-        foreach ($user->getRoles() as $role){
-            
-            if($role == "ROLE_TEACHER") {
+        if($user->getRoles()[0] != "ROLE_TEACHER") {
+            $this->addFlash('warning', 'Vous ne pouvez pas mettre de notes en ligne. Seuls les professeurs le peuvent.');
+        } else {
 
-                if($form->isSubmitted() && $form->isValid()) {
+            if($form->isSubmitted() && $form->isValid()) {
 
-                    $em = $this->getDoctrine()->getManager();
-                    $em->persist($grade);
-                    $em->flush();
-                }
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($grade);
+                $em->flush();
+
+                $this->addFlash('success', 'Note ajoutée avec succès.');
+
             }
         }
 
@@ -60,21 +62,22 @@ class GradeController extends AbstractController
 
         $form->handleRequest($request);
 
-        foreach ($user->getRoles() as $role){
-            
-            if($role == "ROLE_TEACHER") {
+        if($user->getRoles()[0] != "ROLE_TEACHER") {
+            $this->addFlash('warning', 'Vous ne pouvez pas modifier de notes. Seuls les professeurs le peuvent.');
+        } else {
 
-                if($form->isSubmitted() && $form->isValid()) {
+            if($form->isSubmitted() && $form->isValid()) {
 
-                    $em = $this->getDoctrine()->getManager();
-                    $em->flush();
-                }
+                $em = $this->getDoctrine()->getManager();
+                $em->flush();
+
+                $this->addFlash('success', 'Note modifiée avec succès.');
+
             }
         }
 
-
         $formDelete = $this->createForm(DeleteType::class, null, [
-            'action' => $this->generateUrl('grade_delete', ['id' => $id])
+            'action' => $this->generateUrl('grade_delete', ['user_id' => $user_id, 'id' => $id])
         ]);
 
         return $this->render('grade/form.html.twig', [
@@ -85,18 +88,25 @@ class GradeController extends AbstractController
 
 
     /**
-     * @Route("/delete/{id}", name="delete", requirements={"id": "\d+"})
+     * @Route("/user/{user_id}/delete/{id}", name="delete", requirements={"user_id": "\d+","id": "\d+"})
      */
-    public function delete(Grade $grade, Request $request)
+    public function delete($user_id, Grade $grade, Request $request, UserRepository $userRepository)
     {
         $formDelete = $this->createForm(DeleteType::class);
         $formDelete->handleRequest($request);
+        $user = $userRepository->find($user_id);
 
-        if ($formDelete->isSubmitted() && $formDelete->isValid()) {
+        if($user->getRoles()[0] != "ROLE_TEACHER") {
+            $this->addFlash('warning', 'Vous ne pouvez pas supprimer de notes. Seuls les professeurs le peuvent.');
+        } else {
 
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($grade);
-            $em->flush();
+            if ($formDelete->isSubmitted() && $formDelete->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($grade);
+                $em->flush();
+                $this->addFlash('success', 'Note supprimée avec succès.');
+
+            }
         }
 
         return $this->render('grade/form.html.twig', [
